FDG-7268 pulled out desktop-menu.jsx from site-header.jsx
